requirements.txt
aniso8601==7.0.0
cas-client==1.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
mysql-connector-python==8.0.17
packaging==19.0
pluggy==0.12.0
protobuf==3.9.0
py==1.8.0
pycryptodome==3.8.2
PyMySQL==0.9.3
pyparsing==2.4.1.1
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.6
toml==0.10.0
tox==3.13.2
urllib3==1.25.3
virtualenv==16.7.2
Werkzeug==0.15.5
zipp==0.5.2
openpyxl=3.0.3
schedule-0.6.0
